Enable init_sre and importing of the _sre module
fixed a number of small things to get it to work.

Also, updated most of the abort()-ing functions to print out a
Python-level stacktrace before actually calling abort() --
hopefully that won't cause any issues.

+++ src/asm_writing/types.h
@@ -67,9 +67,9 @@ inline bool Register::isCalleeSave() {
 struct Indirect {
 public:
     const Register base;
-    const int offset;
+    const int64_t offset;
 
-    Indirect(const Register base, int offset) : base(base), offset(offset) {}
+    Indirect(const Register base, int64_t offset) : base(base), offset(offset) {}
 };
 
 struct XMMRegister {
